Add cross platform compiler installation script
In order to compile the kernel, we need a compiler for our target
machine. For now we want to target an x86 arch for simplicity.
The script itself is simple and not very flexible but realistically
it will only be used once. Therefore there is very little need to
complicate the process.