Add virtual terminal for the kernel
Because the kernel is compiled without the cstdlib (free-standing)
we are not given any convenient printing functions out of the box.
Furthermore there is no concept of a terminal. The tty module
provides a variety of functions that can manipulate a virtual terminal.
However in its current state it is very limited so don't expect
a full-blown XTERM implementation! The main use is debugging kernel
modules as they are developed.