Disable all GCC optimizations
When optimizations are enabled, it becomes much more difficult to
debug. The user will see the instruction pointer traversing in an odd
manner which will cause confusion. In the future I hope to make the
build step more variable. In any case speed is not the main focus
of this project so its not a top priority.