Prevent attempts to scroll when not possible
Although attempting to scroll up and down when we are at the respective
boundaries results in no change to the scrollback buffer, the
scrollback buffer was re-written to the VGA memory anyway. The current
implementation that copies the scrollback buffer is quite intensive and
should only be used when absolutely necessary.