Add initial C stdlib implementation
Instead of starting a utility project within the kernel it seemed more
practical to leverage a small subset of the C standard library. The
library design closely resembles the actual C standard library but
with some variations such as new functions and function signature
changes. For now it only includes a few functions related to printing.

Note that the Makefile will only compile a libk.a which is only meant
to be linked with kernel objects. User space libc.a will be compilable
in the future.