Isolate boot assembly
In keeping with with general coding conventions, any source code not
related to booting the main kernel code should be in its in own file.
In addition to this change, the Makefile has been updated with one
notable change: the .SECONDARY target. Make decides some files
are intermediate and will delete those files. This was only discovered
by running make with the "-d" switch. Adding a .SECONDARY target with
no prerequisites will force make to preserve all files.