Enable paging on kernel boot
From this point on, the kernel is configured to use IA-32 paging mode.
Specifically the scheme is that of a higher half kernel in which
the codebase is always present within the higher half address space
of every user process. This reduces the need to switch out too many
paging structures like a page directory.

I chose to give the kernel 512 MB to play with because I don't envision
it getting any larger. That means user mode processes get 3.5 GB.
However I tried to make this easily configurable via the boot assembly
file. For now the kernel is configured with with two page tables which
amounts to an 8mb starting resident set.

Segmentation was avoided due to the headache of external fragmentation.